[Feature] sidebar아이콘동작구현
