+++ Cote/Cote/CoteApp.swift
@@ -9,13 +9,12 @@ import SwiftUI
 
 @main
 struct CoteApp: App {
-    @StateObject private var state = UIState()
     var body: some Scene {
         WindowGroup {
             MainView()
                 .toolbar(removing: .title)
                 .toolbarBackgroundVisibility(.hidden, for: .windowToolbar)
-                .environmentObject(state)
+                .environmentObject(UIState())
         }
         .windowResizability(.contentSize)
     }
