+++ Cote/Cote/Presentation/View/Sidebar/SideToolbar.swift
@@ -9,23 +9,35 @@ import SwiftUI
 
 struct SideToolbar: View {
     @State private var selectedButtonID: UUID?
-    
+    @EnvironmentObject private var state: UIState
+
+    private var visibleIcons: [Icon] {
+        if state.isSidebarOpen {
+            return CoteIcon.toolbarIcons
+        } else {
+            return CoteIcon.toolbarIcons.filter { $0.name == "sidebar" }
+        }
+    }
+
     var body: some View {
         HStack(spacing: 4) {
-            
             Spacer()
-            
             HStack(spacing: 0) {
-                ForEach (CoteIcon.toolbarIcons, id: \.id) { button in
-                    MenuButton(selected: Binding(
-                        get: { selectedButtonID == button.id },
-                        set: { if $0 { selectedButtonID = button.id }}),
-                               icon: button
+                ForEach(visibleIcons) { button in
+                    MenuButton(
+                        selected: Binding(
+                            get: { selectedButtonID == button.id },
+                            set: { if $0 { selectedButtonID = button.id } }
+                        ),
+                        icon: button
                     )
+                    .transition(.opacity.combined(with: .scale))
                 }
             }
             .padding(.trailing, 12)
         }
-        .frame(height: 38)
+        //위치고정
+        .frame(width: state.isSidebarOpen ? 210 : 125, height: 38)
+        .animation(.snappy(duration: 0.2), value: state.isSidebarOpen)
     }
 }
